refactor(mcp-server): extract shared helpers from issue-tools and relationship-tools
Move 7 duplicated functions (ensureFieldCache, resolveIssueNodeId,
resolveProjectItemId, updateProjectItemField, getCurrentFieldValue,
resolveConfig, resolveFullConfig) into lib/helpers.ts. Both tool
modules now import from the shared location, eliminating ~400 lines
of duplicated code.
